implemented permission Gates, chart module and httponly cookie login system

+++ app/Http/Controllers/RoleController.php
@@ -2,40 +2,75 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\RoleResource;
 use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
 {
     public function index()
     {
-        return Role::all();
+        Gate::authorize('view', 'roles');
+
+        return RoleResource::collection(Role::all());
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
 
-        return response($role, Response::HTTP_CREATED);
+        if ($permissions = $request->input('permissions')) { //makes sure permissions array is not empty
+            foreach ($permissions as $permission_id) {
+                DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
     public function show($id)
     {
-        return Role::find($id);
+        Gate::authorize('view', 'roles');
+
+        return new RoleResource(Role::find($id));
     }
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::find($id);
 
         $role->update($request->only('name'));
 
-        return response($role, Response::HTTP_ACCEPTED);
+        DB::table('role_permission')->where('role_id', $role->id)->delete();
+
+        if ($permissions = $request->input('permissions')) { //makes sure permissions array is not empty
+            foreach ($permissions as $permission_id) {
+                DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_ACCEPTED);
     }
 
     public function destroy($id)
     {
+        Gate::authorize('edit', 'roles');
+
+        DB::table('role_permission')->where('role_id', $id)->delete();
+
         Role::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
